libs/python-common/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eligibility-common"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "structlog>=24.1",
  "opentelemetry-api>=1.24",
  "opentelemetry-sdk>=1.24",
  "opentelemetry-exporter-otlp>=1.24",
  "opentelemetry-instrumentation-fastapi>=0.45b0",
  "opentelemetry-instrumentation-sqlalchemy>=0.45b0",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "psycopg[binary]>=3.1",
  "redis>=5.0",
  "httpx>=0.27",
  "google-cloud-pubsub>=2.21",
  "tenacity>=8.2",
  "cryptography>=42.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/eligibility_common"]
